src/requirements.txt
aiogram~=2.22.2
aiohttp~=3.8.3
beautifulsoup4~=4.11.1
prettytable~=3.4.1
python-decouple~=3.6